requirements.txt
azure-core==1.33.0
azure-servicebus==7.14.1
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
isodate==0.7.2
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
typing_extensions==4.13.1
urllib3==2.3.0
